pyproject.toml
[project]
name = "ml-oss-framework"
version = "0.1.0"
description = "A clean PyTorch training framework with Hydra configs, AMP, and async checkpointing"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "transformers>=4.30.0",
    "huggingface-hub>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",
    "B008",
    "B905",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "types-pyyaml>=6.0.12.20250915",
]
